Overwrite solve_deps with solve_deps_bis
